services/auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.2